Factor out PrioQueue types
